requirements.txt
pyupbit
pandas
numpy
python-dotenv
requests
tabulate
